feat(*): update code

+++ index.ts
@@ -1,96 +1,94 @@
 /// <reference path="index.d.ts"/>
 
-import './index.scss';
-import { v4 as uuidv4 } from 'uuid';
+import "./index.scss";
+import { v4 as uuidV4 } from "uuid";
 
 /**
  * 常量
  */
 const CONST = {
   CHANNEL: {
-    COPY: 'channel-copy',
-    MOVE: 'channel-move'
+    COPY: "channel-copy",
+    MOVE: "channel-move",
   },
-  CALSS: {
-    BLOCK: 's-block',
-    FRAME: 's-frame'
-  }
-}
+  CLASS: {
+    BLOCK: "s-block",
+    FRAME: "s-frame",
+  },
+};
 
 /**
  * 控件列表元数据
  */
-let blockMetaList: IBlockMata[] = [
+let blockMetaList: IBlockMeta[] = [
   {
-    name: 'block-meta-1',
-    id: uuidv4(),
-    type: 'simple'
+    name: "block-meta-1",
+    id: uuidV4(),
+    type: "simple",
   },
   {
-    name: 'block-meta-2',
-    id: uuidv4(),
-    type: 'simple'
+    name: "block-meta-2",
+    id: uuidV4(),
+    type: "simple",
   },
   {
-    name: 'block-meta-3',
-    id: uuidv4(),
-    type: 'simple'
+    name: "block-meta-3",
+    id: uuidV4(),
+    type: "simple",
   },
   {
-    name: 'block-meta-frame-1',
-    id: uuidv4(),
-    type: 'frame'
+    name: "block-meta-frame-1",
+    id: uuidV4(),
+    type: "frame",
   },
-]
+];
 
 /**
  * 树数据
  */
 let dataTree: IBlock[] = [
   {
-    name: 'block-1',
-    id: uuidv4()
+    name: "block-1",
+    id: uuidV4(),
   },
   {
-    name: 'block-2',
-    id: uuidv4()
+    name: "block-2",
+    id: uuidV4(),
   },
   {
-    name: 'block-frame-20',
-    id: uuidv4(),
+    name: "block-frame-20",
+    id: uuidV4(),
     children: [
       {
-        name: 'block-21',
-        id: uuidv4()
+        name: "block-21",
+        id: uuidV4(),
       },
       {
-        name: 'block-22',
-        id: uuidv4()
+        name: "block-22",
+        id: uuidV4(),
       },
-    ]
+    ],
   },
   {
-    name: 'block-frame-30',
-    id: uuidv4(),
-    children: [
-
-    ]
+    name: "block-frame-30",
+    id: uuidV4(),
+    children: [],
   },
   {
-    name: 'block-3',
-    id: uuidv4()
-  }
-]
+    name: "block-3",
+    id: uuidV4(),
+  },
+];
 
 /**
  * 高亮 dom 数据集
  */
 const oldHightLightData: {
-  $dom?: HTMLElement,
-  type?: TInsertToBlockType
-} = {}
+  $dom?: HTMLElement;
+  type?: TInsertToBlockType;
+} = {};
 
-let currentDraggingId: string = null;
+let currentDraggingId: string = "";
 
 /**
  * 循环1000次，测试性能
@@ -98,40 +96,40 @@ let currentDraggingId: string = null;
 // for (let i = 0; i < 1000; i++) {
 //   dataTree.push({
 //     name: 'block-xxx',
-//     id: uuidv4()
-//   })
+//     id: uuidV4()
+//   })1
 // }
 
 /**
  * 根据 blockMeta 实例化 block
- * @param blockMeta 
- * @returns 
+ * @param blockMeta
+ * @returns
  */
-function makeBlock(blockMeta: IBlockMata): IBlock {
+function makeBlock(blockMeta: IBlockMeta): IBlock {
   const block: IBlock = {
     name: blockMeta.name,
-    id: uuidv4()
-  }
-  if (blockMeta.type === 'frame') {
-    block.children = []
+    id: uuidV4(),
+  };
+  if (blockMeta.type === "frame") {
+    block.children = [];
   }
   return block;
 }
 
-
 /**
  * 从列表中开始拖拽
- * @param ev 
+ * @param ev
  */
 function onMetaDragStart(ev) {
-  const blockMata = blockMetaList.find(blockMata => blockMata.id === ev.target.id)
-  ev.dataTransfer.setData(CONST.CHANNEL.COPY, JSON.stringify(blockMata));
+  const blockMeta = blockMetaList.find(
+    (blockMeta) => blockMeta.id === ev.target.id
+  );
+  ev.dataTransfer.setData(CONST.CHANNEL.COPY, JSON.stringify(blockMeta));
 }
-
 
 /**
  * 落到 container 上
- * @param ev 
+ * @param ev
  */
 function onBlankDragOver(ev) {
   highLightRoot();
@@ -140,7 +138,7 @@ function onBlankDragOver(ev) {
 
 /**
  * 落到 container 上
- * @param ev 
+ * @param ev
  */
 function onBlankDrop(ev) {
   ev.preventDefault();
@@ -149,8 +147,8 @@ function onBlankDrop(ev) {
     const channelMoveStr = ev.dataTransfer.getData(CONST.CHANNEL.MOVE);
 
     if (channelCopyStr) {
-      const blockMata: IBlockMata = JSON.parse(channelCopyStr);
-      insertToRoot(makeBlock(blockMata));
+      const blockMeta: IBlockMeta = JSON.parse(channelCopyStr);
+      insertToRoot(makeBlock(blockMeta));
     } else if (channelMoveStr) {
       const blockId = channelMoveStr;
       const targetBlock = getByID(blockId);
@@ -162,13 +160,13 @@ function onBlankDrop(ev) {
     render();
   } finally {
     clearHightLight();
-    currentDraggingId = null;
+    currentDraggingId = "";
   }
 }
 
 /**
  * 落到树上
- * @param ev 
+ * @param ev
  */
 function onTreeDragOver(ev) {
   highLight(ev);
@@ -178,7 +176,7 @@ function onTreeDragOver(ev) {
 
 /**
  * 从树中开始拖拽
- * @param ev 
+ * @param ev
  */
 function onTreeDragStart(ev) {
   currentDraggingId = ev.target.id;
@@ -190,41 +188,40 @@ function highLightRoot() {
   const $dom = document.getElementById(id);
   if ($dom) {
     // 如果目标元素没有变化，则不执行后续操作
-    if (
-      oldHightLightData.$dom === $dom &&
-      oldHightLightData.type === 'next'
-    ) {
+    if (oldHightLightData.$dom === $dom && oldHightLightData.type === "next") {
       return;
     }
 
     clearHightLight();
 
     oldHightLightData.$dom = $dom;
-    oldHightLightData.type = 'next';
+    oldHightLightData.type = "next";
 
-    $dom.style.borderBottom = '3px solid blue';
-    $dom.style.marginBottom = '-2px';
+    $dom.style.borderBottom = "3px solid blue";
+    $dom.style.marginBottom = "-2px";
   }
 }
 
 /**
  * 拖拽区域样式高亮
- * 
- * @param ev 
+ *
+ * @param ev
  */
 function highLight(ev) {
-  const [, type, $wrap] = findTargetDropDom(ev);
+  const target = findTargetDropDom(ev);
+  if (!target) {
+    return;
+  }
+
+  const [, type, $wrap] = target;
 
   let $dom = $wrap;
-  if (type === 'before' || type == 'next') {
+  if (type === "before" || type == "next") {
     $dom = $wrap.children[0] as HTMLElement;
   }
 
   // 如果目标元素没有变化，则不执行后续操作
-  if (
-    oldHightLightData.$dom === $dom &&
-    oldHightLightData.type === type
-  ) {
+  if (oldHightLightData.$dom === $dom && oldHightLightData.type === type) {
     return;
   }
 
@@ -233,32 +230,31 @@ function highLight(ev) {
   oldHightLightData.$dom = $dom;
   oldHightLightData.type = type;
 
-  if (type === 'inner') {
-    $dom.style.background = '#40abff';
-  } else if (type === 'before') {
-    $dom.style.borderTop = '3px solid blue';
-    $dom.style.marginTop = '-2px';
-  } else if (type === 'next') {
-    $dom.style.borderBottom = '3px solid blue';
-    $dom.style.marginBottom = '-2px';
+  if (type === "inner") {
+    $dom.style.background = "#40abff";
+  } else if (type === "before") {
+    $dom.style.borderTop = "3px solid blue";
+    $dom.style.marginTop = "-2px";
+  } else if (type === "next") {
+    $dom.style.borderBottom = "3px solid blue";
+    $dom.style.marginBottom = "-2px";
   }
 }
 
 function clearHightLight() {
   // 清除老的高亮样式
   if (oldHightLightData.$dom) {
-    oldHightLightData.$dom.style.background = '';
-    oldHightLightData.$dom.style.borderTop = '';
-    oldHightLightData.$dom.style.borderBottom = '';
-    oldHightLightData.$dom.style.marginTop = '';
-    oldHightLightData.$dom.style.marginBottom = '';
+    oldHightLightData.$dom.style.background = "";
+    oldHightLightData.$dom.style.borderTop = "";
+    oldHightLightData.$dom.style.borderBottom = "";
+    oldHightLightData.$dom.style.marginTop = "";
+    oldHightLightData.$dom.style.marginBottom = "";
   }
 }
 
-
 /**
  * 落到树上
- * @param ev 
+ * @param ev
  */
 function onTreeDrop(ev) {
   try {
@@ -266,97 +262,113 @@ function onTreeDrop(ev) {
     ev.stopPropagation();
     const channelCopyStr = ev.dataTransfer.getData(CONST.CHANNEL.COPY);
     const channelMoveStr = ev.dataTransfer.getData(CONST.CHANNEL.MOVE);
-    const [id, type] = findTargetDropDom(ev);
+    const target = findTargetDropDom(ev);
+    if (!target) {
+      return;
+    }
+
+    const [id, type] = target;
 
     if (channelCopyStr) {
-      const blockMata: IBlockMata = JSON.parse(channelCopyStr);
-      insertToBlock(getByID(id), makeBlock(blockMata), type)
+      const blockMeta: IBlockMeta = JSON.parse(channelCopyStr);
+      insertToBlock(getByID(id), makeBlock(blockMeta), type);
     } else if (channelMoveStr) {
       const blockId = channelMoveStr;
       // 如果目标 ID 和 要移动的 ID 相同，则表示无需移动
       if (blockId !== id) {
         const _block = getByID(blockId);
         deleteByID(dataTree, blockId);
-        insertToBlock(getByID(id), _block, type)
+        insertToBlock(getByID(id), _block, type);
       }
     }
     render();
   } finally {
     clearHightLight();
-    currentDraggingId = null;
+    currentDraggingId = "";
   }
 }
 
 /**
  * 从 event 判断需要落到哪个drop区域
- * 
- * @param ev 
+ *
+ * @param ev
  * @returns [对应控件的ID，放入的类型， 目标Dom]
  */
-function findTargetDropDom(ev): [string, TInsertToBlockType, HTMLElement] {
-  let path: HTMLElement[] = ev.path;
-  // 下面逻辑用于处理 当前block 落在自己 子block 的情况 
+function findTargetDropDom(
+  ev
+): [string, TInsertToBlockType, HTMLElement] | undefined {
+  let path: HTMLElement[] = ev.composedPath();
+  // 下面逻辑用于处理 当前block 落在自己 子block 的情况
   // === start ===
   const $dragDom = document.getElementById(currentDraggingId);
-  const index = path.findIndex(i => i === $dragDom) + 1;
+  const index = path.findIndex((i) => i === $dragDom) + 1;
   path = path.slice(index);
   // === end ===
 
-  let type: TInsertToBlockType;
+  let type: TInsertToBlockType | undefined;
+
   for (let x = 0; x < path.length; x++) {
     if (path[x].classList !== undefined) {
       const classList = [...path[x].classList];
       for (let y = 0; y < classList.length; y++) {
-        if (classList[y] === CONST.CALSS.FRAME) {
-          type = 'inner'
+        if (classList[y] === CONST.CLASS.FRAME) {
+          type = "inner";
         }
-        if (classList[y] === CONST.CALSS.BLOCK) {
+        if (classList[y] === CONST.CLASS.BLOCK) {
           const { y } = ev;
           const { height, top } = path[x].getBoundingClientRect();
           if (height > (y - top) * 2) {
-            type = 'before'
+            type = "before";
           } else {
-            type = 'next'
+            type = "next";
           }
         }
         if (type) {
-          return [path[x].dataset.id, type, path[x]];
+          return [path[x].dataset.id || "", type, path[x]];
         }
       }
     }
   }
 }
 
-
 /**
  * 渲染容器控件树
  */
 function render() {
-  const $container = document.querySelector('#container');
+  const $container = document.querySelector("#container");
+  if (!$container) {
+    return;
+  }
   const _render = () => {
-    const _genblock = (item, isLast) => {
+    const _genBlock = (item, isLast) => {
       const { id, name, children } = item;
-      let childrenStr = '';
+      let childrenStr = "";
       if (children) {
         childrenStr = `
-          <div class="block-frame ${CONST.CALSS.FRAME}" data-id='${id}'>
-            ${children.map((i, index) => _genblock(i, index === children.length - 1)).join('')}
+          <div class="block-frame ${CONST.CLASS.FRAME}" data-id='${id}'>
+            ${children
+              .map((i, index) => _genBlock(i, index === children.length - 1))
+              .join("")}
           </div>
-        `
+        `;
       }
       return `
-        <div class="block-wrap ${CONST.CALSS.BLOCK}"  data-id='${id}' ondrop="onTreeDrop(event)" ondragover="onTreeDragOver(event)">
+        <div class="block-wrap ${
+          CONST.CLASS.BLOCK
+        }"  data-id='${id}' ondrop="onTreeDrop(event)" ondragover="onTreeDragOver(event)">
           <div id='${id}' class="block" draggable="true" ondragstart="onTreeDragStart(event)">
             <div class="title">${name}</div>
             ${childrenStr}
           </div>
-          ${isLast ? '' : '<div class="block-arrow"></div>'}
+          ${isLast ? "" : '<div class="block-arrow"></div>'}
         </div>
-      `
-    }
-    return dataTree.map((i, index) => _genblock(i, index === dataTree.length - 1)).join('')
-  }
-  // render 后去除对 highLightDome 的指针引用
+      `;
+    };
+    return dataTree
+      .map((i, index) => _genBlock(i, index === dataTree.length - 1))
+      .join("");
+  };
+  // render 后去除对 highLightDom 的指针引用
   oldHightLightData.$dom = undefined;
   oldHightLightData.type = undefined;
   $container.innerHTML = _render();
@@ -366,19 +378,24 @@ function render() {
  * 渲染左侧控件列表
  */
 function renderMeta() {
-  const $blockMetaList = document.querySelector('#block-meta-list')
-  const str = blockMetaList.map(blockMeta => {
-    return ` <div draggable="true" id="${blockMeta.id}" class="block-meta" ondragstart="onMetaDragStart(event)">${blockMeta.name}</div>`
-  }).join('')
+  const $blockMetaList = document.querySelector("#block-meta-list");
+  if (!$blockMetaList) {
+    return;
+  }
+  const str = blockMetaList
+    .map((blockMeta) => {
+      return ` <div draggable="true" id="${blockMeta.id}" class="block-meta" ondragstart="onMetaDragStart(event)">${blockMeta.name}</div>`;
+    })
+    .join("");
   $blockMetaList.innerHTML = str;
 }
 
 /**
  * 控件添加到 根
- * @param block 
+ * @param block
  */
-function insertToRoot(block: IBlock, type: TInsertToRoot = 'next') {
-  if (type === 'next') {
+function insertToRoot(block: IBlock, type: TInsertToRoot = "next") {
+  if (type === "next") {
     dataTree.push(block);
   } else {
     dataTree.unshift(block);
@@ -388,20 +405,24 @@ function insertToRoot(block: IBlock, type: TInsertToRoot = 'next') {
 
 /**
  * 基于某个 block 为锚点添加 block
- * @param target 
- * @param block 
- * @param type 
+ * @param target
+ * @param block
+ * @param type
  */
-function insertToBlock(target: IBlock, block: IBlock, type: TInsertToBlockType = 'next') {
-  if (!target) throw 'target is required; fn insertToBlock';
-  if (!block) throw 'block is required; fn insertToBlock';
-  if (type === 'next') {
+function insertToBlock(
+  target: IBlock,
+  block: IBlock,
+  type: TInsertToBlockType = "next"
+) {
+  if (!target) throw "target is required; fn insertToBlock";
+  if (!block) throw "block is required; fn insertToBlock";
+  if (type === "next") {
     const list = target.parent?.children || dataTree;
-    const index = list.findIndex(block => block.id === target.id);
+    const index = list.findIndex((block) => block.id === target.id);
     list.splice(index + 1, 0, block);
-  } else if (type === 'before') {
+  } else if (type === "before") {
     const list = target.parent?.children || dataTree;
-    const index = list.findIndex(block => block.id === target.id);
+    const index = list.findIndex((block) => block.id === target.id);
     list.splice(index, 0, block);
   } else {
     if (target.children) {
@@ -411,12 +432,11 @@ function insertToBlock(target: IBlock, block: IBlock, type: TInsertToBlockType =
   link();
 }
 
-
 /**
  * 根据 id 删除树中的 block
- * @param arr 
- * @param id 
- * @returns 
+ * @param arr
+ * @param id
+ * @returns
  */
 function deleteByID(arr: IBlock[], id: string) {
   for (let x = arr.length - 1; x >= 0; x--) {
@@ -434,13 +454,13 @@ function deleteByID(arr: IBlock[], id: string) {
 
 /**
  * 根据 id 获得树中的 block
- * @param id 
- * @returns 
+ * @param id
+ * @returns
  */
 function getByID(id: string) {
   let target;
   const _getByID = (arr: IBlock[], id: string) => {
-    arr.forEach(block => {
+    arr.forEach((block) => {
       if (block.id === id) {
         target = block;
         return;
@@ -448,21 +468,21 @@ function getByID(id: string) {
       if (block.children) {
         _getByID(block.children, id);
       }
-    })
-  }
+    });
+  };
   _getByID(dataTree, id);
   return target;
 }
 
 function link() {
   const _link = (data: IBlock[], parent?: IBlock) => {
-    data.forEach(block => {
+    data.forEach((block) => {
       block.parent = parent;
       if (block.children) {
-        _link(block.children, block)
+        _link(block.children, block);
       }
-    })
-  }
+    });
+  };
   _link(dataTree);
 }
 
@@ -473,7 +493,6 @@ function init() {
 }
 
 init();
-
 
 (window as any).onMetaDragStart = onMetaDragStart;
